fix: clicking on label whitespace no longer focuses input
Clicking on the label itself still works.

BREAKING CHANGE: The label whitespace fix is a breaking change because it changes the layout of the label.
Style top no longer automatically does w-full. You must place an additional <div class="flex-1"/> after the label.
The top style will probably be removed in the future.