feat: dropped fontawesome in favor of unplugin-icons
This also fixes icons not working unless fontawesome's icons are registered due
to the different way unplugin-icons works.

<!--skip-release-notes-->
BREAKING CHANGES: The Fa fontawesome component has been removed.

- There is now an icon component but it is not for "picking" an icon, just a
wrapper to make it have the right height when it's alone in a div.

- Instructions have been added for how to add unplugin to a downstream project
to use it similarly to how it's used internally.

- Any components that took in an icon prop no longer do.

- The button component no longer has an icon slot nested inside the label slot,
but separate icon, iconAfter, and label slots.