feat: improved suggestions
The suggestions logic is now de-coupled from the component and can be re-used if
needed.

- The submit event also works more consistently.

- The open/close logic in general has been cleaned up.

- Added a `suggestionSelector` prop option for determining which option to
select (this can be used to implement fuzzy matching).

- You can now filter the handling of all events to the suggestions component.

BREAKING CHANGES:

- `isValid`, `isOpen` and `activeSuggestion` are no longer v-models. `isValid`
is only a prop, the others are now only events.

- The suggestions component should be controled instead from the `suggestions`
object it exposes which contains all the variables and functions from the
composable. Before these were partially exposed at the root.

- The component is no longer generic for the model or input value only for the
suggestions list and it still has some edge cases to iron out. In general though
it works better than before.

- `data-is-active-suggestion` has been removed, given aria-selected is the same
thing.

- The way the submit event works is now slightly different.