build: fixed playground
Couldn;t get tailwind to work in the playground because the @witchcraft/ui
module it tries to import doesn't technically exist (unless working with my
local monorepo). Using absolute paths for the style/base imports doesn't work
either.

Instead we link to the parent folder before building. This should work
regardless of the environment used to build the package.