v0breaking: added useDialogForBackdrop option to the Popup component and defaulted it to false
There are now two options, useBackdrop (which is only about whether to use a
backdrop or not, where before if true, the backdrop was always a dialog
element), and useDialogForBackdrop, which will use a dialog element for the
backdrop. This now defaults to false.

While it would be nice to use the native dialog for this, it requires additional
setup if using css variables for theming as variables must be set on :dialog or
:root (the parent component is not enough). This is highly inconvenient as it's
nice to be able to encapsulate the entire theme around a section of an app by
registering variables on a component. Because of this, using a dialog,
components can seem to break in weird ways. Most styles will work, but not
custom variables set by the theming library.