v0feat(table): added virtualization to table
This involves some minor breaking changes. Note virtualization is NOT enabled by
default.

- (breaking) The table is now wrapped in two divs. The class attribute is still
forwarded to the table, but now wrapperClass can be used to forward classes to
the root wrapper div. This was required to make virtualization work.
.resizable-cols-setup is now also added to the wrapper.

- If using slots with virtualization be sure to read the docs for the
virtualizerOptions and update the props passed to your slot element.

- (breaking) cells are now styled as truncate by default to avoid issues with
virtualization.

- Improved performance not just via virtualization: Classes are now all
pre-computed including style merges, instead of inline in the template.

- Added stories to testvirtualization.

- Added has-scrollbar-* classes so we can style overflowed states better

- This required several tricks to get working with sticky headers and fit width
false and resize in general. Read docs carefully.