v0fix(table): fixed some more resize issues
- It was still possible in some circumstances for columns to go below the min
margin width. That's no longer the case, when elements reach the min margin
width, their widths are now set in pixels and the table will scroll, even with
fitWidth true.

- There was some jitter when the container started to scroll along the x axis.
With the scrollbar appearing/dissapearing. This was due to a previous faulty fix
which didn't correctly address the underlying issues (the above min width
problem).